feat: rename keys.rs to key.rs
