+++ src/meta-srv/src/service/admin/node_lease.rs
@@ -20,7 +20,7 @@ use tonic::codegen::http;
 
 use crate::cluster::MetaPeerClientRef;
 use crate::error::{self, Result};
-use crate::keys::{LeaseKey, LeaseValue};
+use crate::key::{LeaseKey, LeaseValue};
 use crate::lease;
 use crate::service::admin::{util, HttpHandler};
